fix: optimized imports and change the userId type to solve some errors

+++ app/src/main/java/com/task/management/workflow/calendar/domain/CreateEventRequest.kt
@@ -1,10 +1,8 @@
 package com.task.management.workflow.calendar.domain
 
-import java.time.LocalDate
-
 class CreateEventRequest(
     val projectId: Long,
-    val userId: Long,
+    val userId: Int,
     val title: String,
     val description: String,
     var dueDate: String
